requirements.txt
fastapi
uvicorn[standard]
anthropic
python-docx
pydantic
python-dotenv
httpx
pillow
python-multipart
